ml-offline/requirements.txt
pandas==2.2.0
scikit-learn==1.4.0
xgboost==2.0.3
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.1
